+++ Rag_5040/NIN.lua
@@ -163,6 +163,8 @@ local WeakElementTable = {
 gcmelee = gFunc.LoadFile('common\\gcmelee.lua')
 
 profile.HandleAbility = function()
+    gcmelee.DoAbility()
+
     gFunc.EquipSet(sets.Hate)
 end