+++ Rag_5040/WHM.lua
@@ -9,8 +9,6 @@ local blmSJMaxMP = nil -- The Max MP you have when /blm in your idle set
 local virology_ring = true
 local virology_ring_slot = 'Ring2'
 
-local displayheadOnAbility = true
-
 local sets = {
     Idle = {},
     IdleALT = {},
@@ -95,9 +93,7 @@ Everything below can be ignored.
 gcmage = gFunc.LoadFile('common\\gcmage.lua')
 
 profile.HandleAbility = function()
-    if (displayheadOnAbility) then
-        AshitaCore:GetChatManager():QueueCommand(-1, '/displayhead')
-    end
+    gcmage.DoAbility()
 end
 
 profile.HandleItem = function()